SDDK folder: split more files into hpp / cpp
