+++ src/SDDK/communicator.hpp
@@ -95,7 +95,8 @@ struct mpi_type_wrapper<double>
     }
 };
 
-template <>
+
+    template <>
 struct mpi_type_wrapper<long double>
 {
     static MPI_Datatype kind()
@@ -794,71 +795,9 @@ class Communicator
 };
 
 /// Get number of ranks per node.
-inline int num_ranks_per_node()
-{
-    static int num_ranks{-1};
-    if (num_ranks == -1) {
-        char name[MPI_MAX_PROCESSOR_NAME];
-        int len;
-        CALL_MPI(MPI_Get_processor_name, (name, &len));
-        std::vector<size_t> hash(Communicator::world().size());
-        hash[Communicator::world().rank()] = std::hash<std::string>{}(std::string(name, len));
-        Communicator::world().allgather(hash.data(), Communicator::world().rank(), 1);
-        std::sort(hash.begin(), hash.end());
+int num_ranks_per_node();
 
-        int n{1};
-        for (int i = 1; i < (int)hash.size(); i++) {
-            if (hash[i] == hash.front()) {
-                n++;
-            } else {
-                break;
-            }
-        }
-        int m{1};
-        for (int i = (int)hash.size() - 2; i >= 0; i--) {
-            if (hash[i] == hash.back()) {
-                m++;
-            } else {
-                break;
-            }
-        }
-        num_ranks = std::max(n, m);
-    }
-
-    return num_ranks;
-}
-
-inline int get_device_id(int num_devices__)
-{
-    static int id{-1};
-    if (num_devices__ == 0) {
-        return id;
-    }
-    if (id == -1) {
-        #pragma omp single
-        {
-            int r = Communicator::world().rank();
-            char name[MPI_MAX_PROCESSOR_NAME];
-            int len;
-            CALL_MPI(MPI_Get_processor_name, (name, &len));
-            std::vector<size_t> hash(Communicator::world().size());
-            hash[r] = std::hash<std::string>{}(std::string(name, len));
-            Communicator::world().allgather(hash.data(), r, 1);
-            std::map<size_t, std::vector<int>> rank_map;
-            for (int i = 0; i < Communicator::world().size(); i++) {
-                rank_map[hash[i]].push_back(i);
-            }
-            for (int i = 0; i < (int)rank_map[hash[r]].size(); i++) {
-                if (rank_map[hash[r]][i] == r) {
-                    id = i % num_devices__;
-                    break;
-                }
-            }
-        }
-        assert(id >= 0);
-    }
-    return id;
-}
+int get_device_id(int num_devices__);
 
 /// Parallel standard output.
 /** Proveides an ordered standard output from multiple MPI ranks. */
@@ -882,49 +821,9 @@ class pstdout
         flush();
     }
 
-    void printf(const char* fmt, ...)
-    {
-        std::vector<char> str(1024); // assume that one printf will not output more than this
+    void printf(const char* fmt, ...);
 
-        std::va_list arg;
-        va_start(arg, fmt);
-        int n = vsnprintf(&str[0], str.size(), fmt, arg);
-        va_end(arg);
-
-        n = std::min(n, (int)str.size());
-
-        if ((int)buffer_.size() - count_ < n) {
-            buffer_.resize(buffer_.size() + str.size());
-        }
-        std::memcpy(&buffer_[count_], &str[0], n);
-        count_ += n;
-    }
-
-    void flush()
-    {
-        std::vector<int> counts(comm_.size());
-        comm_.allgather(&count_, counts.data(), comm_.rank(), 1);
-
-        int offset{0};
-        for (int i = 0; i < comm_.rank(); i++) {
-            offset += counts[i];
-        }
-
-        /* total size of the output buffer */
-        int sz = count_;
-        comm_.allreduce(&sz, 1);
-
-        if (sz != 0) {
-            std::vector<char> outb(sz + 1);
-            comm_.allgather(&buffer_[0], &outb[0], offset, count_);
-            outb[sz] = 0;
-
-            if (comm_.rank() == 0) {
-                std::printf("%s", &outb[0]);
-            }
-        }
-        count_ = 0;
-    }
+    void flush();
 };
 
 } // namespace sddk
